Refactor gameController and displayController
Removed hard coded create player variables.
Connected DOM input field text to the gameController to create players
for the game.
Selected DOM elements relevant to the game.
Add event listener to start button to start the game.
Disable game when a winner is found or game ends in a tie.
Reset button resets the tiles for another round but maintains the score